requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
geopy==2.3.0
shapely==2.0.2
numpy<2 